pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "ai-agents-memory"
version = "0.1.0"
description = "Runnable code samples for LLM agent memory management techniques."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "AI Agents Memory Contributors" }]
dependencies = [
  "chromadb>=0.5.0",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "langgraph>=0.2.0",
  "llama-index-core>=0.12.0",
  "llama-index-embeddings-openai>=0.3.0",
  "llama-index-llms-openai>=0.3.0",
  "llama-index-vector-stores-chroma>=0.4.0",
  "pydantic>=2.8.0",
  "rich>=13.7.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
memory-samples = "ai_agents_memory.cli:app"

[tool.pytest.ini_options]
addopts = "-q"
markers = [
  "requires_openai: integration tests that call the OpenAI API and may incur cost",
]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_agents_memory"]
